pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.4.6",
    "arxiv>=2.2.0",
    "docker>=7.1.0",
    "duckduckgo-search>=8.0.1",
    "fastapi[standard]>=0.115.12",
    "fastmcp>=2.3.3",
    "httpx>=0.28.1",
    "lxml-html-clean>=0.4.2",
    "nest-asyncio>=1.6.0",
    "newspaper4k>=0.9.3.1",
    "openai>=1.78.1",
    "paramiko>=3.5.1",
    "pydantic>=2.11.4",
    "pypdf>=5.5.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.40",
    "tavily-python>=0.7.2",
    "tzlocal>=5.3.1",
    "uvicorn>=0.34.2",
    "yfinance>=0.2.61",
]
